pyproject.toml
[project]
name = "quarry"
version = "0.1.0"
description = "Agentic job search system"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13",
  "boto3>=1.34.0",
  "requests>=2.31.0",
  "apscheduler>=3.10.0",
  "flask>=3.0.0",
  "tenacity>=8.2.0",
  "pyyaml>=6.0",
  "sentence-transformers>=2.7.0",
  "scikit-learn>=1.4.0",
  "numpy>=1.26.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",
  "python-dotenv>=1.0.0",
  "click>=8.1.0",
  "tabulate>=0.9.0",
  "pandas>=2.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "geonamescache>=3.0.0",
  "python-jobspy @ git+https://github.com/cullenwatson/JobSpy.git",
  "sqlalchemy>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff", "pyright", "pre-commit"]

[tool.setuptools.packages.find]
include = ["quarry*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = "."

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.11"
